+++ src/lib/filters.ts
@@ -1,5 +1,5 @@
-import { PR } from '../state/types';
-import { TabOptions, FilterTypes } from '../tabs/TabTypes';
+import { FilterTypes } from '../tabs/TabTypes';
+import { PR, TabOptions } from '../state/types';
 
 export type FilterFunc = (pullRequest: PR, filterValue: string[]) => boolean;
 
@@ -14,7 +14,8 @@ export const filterByTitle: FilterFunc = (pullRequest, filterValue) =>
   pullRequest.pullRequestId.toString().indexOf(filterValue[0].toLocaleLowerCase()) > -1 ||
   pullRequest.repository.name.toLocaleLowerCase().indexOf(filterValue[0].toLocaleLowerCase()) > -1 ||
   pullRequest.sourceBranch.name.toLocaleLowerCase().indexOf(filterValue[0].toLocaleLowerCase()) > -1 ||
-  pullRequest.targetBranch.name.toLocaleLowerCase().indexOf(filterValue[0].toLocaleLowerCase()) > -1;
+  pullRequest.targetBranch.name.toLocaleLowerCase().indexOf(filterValue[0].toLocaleLowerCase()) > -1 ||
+  pullRequest.createdBy.displayName.toLocaleLowerCase().indexOf(filterValue[0].toLocaleLowerCase()) > -1;
 
 export const filterByRepositoryId: FilterFunc = (pullRequest, filterValue) =>
   filterValue.some(v => pullRequest.repositoryId === v);
@@ -34,8 +35,11 @@ export const filterByReviewers: FilterFunc = (pullRequest, filterValue) =>
 export const filterByApprovalStatus: FilterFunc = (pullRequest, filterValue) =>
   pullRequest.myApprovalStatus.toString() === filterValue[0].toString();
 
-export const filterByDraftStatus: FilterFunc = (pullRequest, filterValue) =>
-  pullRequest.isDraft === (filterValue[0] === 'true');
+export const filterByDraftStatus: FilterFunc = pullRequest => pullRequest.isDraft;
+
+export const filterByActiveStatus: FilterFunc = pullRequest => pullRequest.isActive;
+
+export const filterByCompletedStatus: FilterFunc = pullRequest => pullRequest.isCompleted;
 
 export const setupFilters = (filterValues: Partial<Record<FilterTypes, any>>, tab: TabOptions) => {
   const { searchString, repositories, sourceBranch, targetBranch, author, reviewer, myApprovalStatus } = filterValues;
@@ -77,8 +81,8 @@ export const setupFilters = (filterValues: Partial<Record<FilterTypes, any>>, ta
         isActive: myApprovalStatus !== undefined && myApprovalStatus.length > 0,
       },
       {
-        func: filterByDraftStatus,
-        val: ['false'],
+        func: filterByActiveStatus,
+        val: undefined,
         isActive: true,
       },
     ],
@@ -120,7 +124,49 @@ export const setupFilters = (filterValues: Partial<Record<FilterTypes, any>>, ta
       },
       {
         func: filterByDraftStatus,
-        val: ['true'],
+        val: undefined,
+        isActive: true,
+      },
+    ],
+    recentlyCompleted: [
+      {
+        func: filterByTitle,
+        val: [searchString || ''],
+        isActive: searchString !== undefined && searchString.length > 0 && searchString[0].length > 0,
+      },
+      {
+        func: filterByRepositoryId,
+        val: repositories || [],
+        isActive: repositories !== undefined && repositories.length > 0,
+      },
+      {
+        func: filterBySourceBranchDisplayName,
+        val: sourceBranch || [],
+        isActive: sourceBranch !== undefined && sourceBranch.length > 0,
+      },
+      {
+        func: filterByTargetBranchDisplayName,
+        val: targetBranch || [],
+        isActive: targetBranch !== undefined && targetBranch.length > 0,
+      },
+      {
+        func: filterByCreatedByUserId,
+        val: author || [],
+        isActive: author !== undefined && author.length > 0,
+      },
+      {
+        func: filterByReviewers,
+        val: reviewer || [],
+        isActive: reviewer !== undefined && reviewer.length > 0,
+      },
+      {
+        func: filterByApprovalStatus,
+        val: myApprovalStatus || [],
+        isActive: myApprovalStatus !== undefined && myApprovalStatus.length > 0,
+      },
+      {
+        func: filterByCompletedStatus,
+        val: undefined,
         isActive: true,
       },
     ],