+++ src/state/actions.ts
@@ -1,10 +1,9 @@
 import { Action, Dispatch } from 'redux';
 
 // azure-devops-extension-sdk
-import * as DevOps from 'azure-devops-extension-sdk';
+import { getService, getUser, getExtensionContext, getAccessToken } from 'azure-devops-extension-sdk';
 
 // azure-devops-extension-api
-import { ResourceRef } from 'azure-devops-extension-api/WebApi/WebApi';
 import { GitRestClient } from 'azure-devops-extension-api/Git/GitClient';
 import {
   getClient,
@@ -13,22 +12,19 @@ import {
   IExtensionDataService,
   IExtensionDataManager,
 } from 'azure-devops-extension-api';
+import { GitPullRequestSearchCriteria, PullRequestStatus } from 'azure-devops-extension-api/Git/Git';
 import { WorkItemTrackingRestClient } from 'azure-devops-extension-api/WorkItemTracking/WorkItemTrackingClient';
 
-import { initialState } from './store';
 import { FilterDictionary } from '../tabs/TabTypes';
-import { sortByRepositoryName, sortByPullRequestId } from '../lib/utils';
 import { fromPullRequestToPR } from './transformData';
-import { ActionTypes, SavedPrHubState } from './types';
-import { GitPullRequest, GitPullRequestSearchCriteria, PullRequestStatus } from 'azure-devops-extension-api/Git/Git';
+import { ActionTypes, DefaultSettings, SortDirection } from './types';
+import { sortByRepositoryName, sortByPullRequestId } from '../lib/utils';
 
-// action interfaces
 export interface FetchAction extends Action {
   payload?: any;
 }
 
-// criteria setup
-export const pullRequestCriteria: GitPullRequestSearchCriteria = {
+export const activePrCriteria: GitPullRequestSearchCriteria = {
   repositoryId: '',
   creatorId: '',
   includeLinks: true,
@@ -39,37 +35,51 @@ export const pullRequestCriteria: GitPullRequestSearchCriteria = {
   targetRefName: '',
 };
 
+export const completedPrCriteria: GitPullRequestSearchCriteria = {
+  repositoryId: '',
+  creatorId: '',
+  includeLinks: true,
+  reviewerId: '',
+  sourceRefName: '',
+  sourceRepositoryId: '',
+  status: PullRequestStatus.Completed,
+  targetRefName: '',
+};
+
 // VSTS REST Clients
 export const gitClient: GitRestClient = getClient(GitRestClient);
 export const workItemClient: WorkItemTrackingRestClient = getClient(WorkItemTrackingRestClient);
 
 const getRepositories = async () => {
-  const projectService = await DevOps.getService<IProjectPageService>('ms.vss-tfs-web.tfs-page-data-service');
+  const projectService = await getService<IProjectPageService>('ms.vss-tfs-web.tfs-page-data-service');
   const currentProject = await projectService.getProject();
   return (await gitClient.getRepositories(currentProject?.name, true)).sort(sortByRepositoryName);
 };
 
-const getWorkItemsForPr = async (pullRequest: GitPullRequest) => {
-  const workItemRefs = await gitClient.getPullRequestWorkItemRefs(pullRequest.repository.id, pullRequest.pullRequestId);
-  const workItemIds = workItemRefs.flatMap((ref: ResourceRef) => Number(ref.id));
-  return workItemIds.length > 0 ? await workItemClient.getWorkItems(workItemIds) : [];
+// const getWorkItemsForPr = async (pullRequest: GitPullRequest) => {
+//   const workItemRefs = await gitClient.getPullRequestWorkItemRefs(pullRequest.repository.id, pullRequest.pullRequestId);
+//   const workItemIds = workItemRefs.flatMap((ref: ResourceRef) => Number(ref.id));
+//   return workItemIds.length > 0 ? await workItemClient.getWorkItems(workItemIds) : [];
+// };
+
+const getLayoutService = async (): Promise<IHostPageLayoutService> => {
+  return await getService<IHostPageLayoutService>('ms.vss-features.host-page-layout-service');
 };
 
-const setFullScreenMode = async (): Promise<boolean> => {
-  const layoutService = await DevOps.getService<IHostPageLayoutService>('ms.vss-features.host-page-layout-service');
-  const fullScreenMode = await layoutService.getFullScreenMode();
-  layoutService.setFullScreenMode(!fullScreenMode);
-  return !fullScreenMode;
+const setFullScreenModeState = async (isFullScreenMode: boolean): Promise<boolean> => {
+  const layoutService = await getLayoutService();
+  layoutService.setFullScreenMode(isFullScreenMode);
+  return isFullScreenMode;
 };
 
-/**
- * @summary Synchronously sets current user in redux store
- */
 export const setCurrentUser = () => (dispatch: Dispatch<FetchAction>) =>
-  dispatch({ type: ActionTypes.SET_CURRENT_USER, payload: DevOps.getUser() });
+  dispatch({ type: ActionTypes.SET_CURRENT_USER, payload: getUser() });
 
 export const toggleSortDirection = () => (dispatch: Dispatch<FetchAction>) =>
   dispatch({ type: ActionTypes.TOGGLE_SORT_DIRECTION });
+
+export const setSortDirection = (sortDirection: SortDirection) => (dispatch: Dispatch<FetchAction>) =>
+  dispatch({ type: ActionTypes.SET_SORT_DIRECTION, payload: sortDirection });
 
 export const triggerSortDirection = () => (dispatch: Dispatch<FetchAction>) =>
   dispatch({ type: ActionTypes.TRIGGER_SORT_DIRECTION });
@@ -79,6 +89,9 @@ export const setFilterValues = (filterValues: FilterDictionary) => (dispatch: Di
 
 export const setSelectedTab = (newSelectedTab: string) => (dispatch: Dispatch<FetchAction>) =>
   dispatch({ type: ActionTypes.SET_SELECTED_TAB, payload: newSelectedTab });
+
+export const setFilterBar = (isFilterVisible: boolean) => (dispatch: Dispatch<FetchAction>) =>
+  dispatch({ type: ActionTypes.SET_FILTER_BAR, payload: isFilterVisible });
 
 export const toggleSettingsPanel = () => (dispatch: Dispatch<FetchAction>) =>
   dispatch({ type: ActionTypes.TOGGLE_SETTINGS_PANEL });
@@ -93,7 +106,12 @@ export const setPullRequests = () => async (dispatch: Dispatch<FetchAction>) => 
   dispatch({ type: ActionTypes.ADD_ASYNC_TASK });
   const repositories = await getRepositories();
   const getAllRepositoryPullRequests = repositories.map(
-    async repo => await gitClient.getPullRequests(repo.id, pullRequestCriteria),
+    async repo => await gitClient.getPullRequests(repo.id, activePrCriteria),
+  );
+  getAllRepositoryPullRequests.push(
+    ...repositories.map(
+      async repo => await gitClient.getPullRequests(repo.id, completedPrCriteria, undefined, undefined, undefined, 10),
+    ),
   );
   const allRepositoryPullRequests = await Promise.all(getAllRepositoryPullRequests);
   const getCompletePullRequests = allRepositoryPullRequests.flatMap(prsForSingleRepo =>
@@ -104,8 +122,9 @@ export const setPullRequests = () => async (dispatch: Dispatch<FetchAction>) => 
     allPullRequests.map(async pullRequest =>
       fromPullRequestToPR({
         pr: pullRequest,
-        workItems: await getWorkItemsForPr(pullRequest),
-        userContext: DevOps.getUser(),
+        workItems: [],
+        // workItems: await getWorkItemsForPr(pullRequest),
+        userContext: getUser(),
       }),
     ),
   );
@@ -119,9 +138,6 @@ export const setPullRequests = () => async (dispatch: Dispatch<FetchAction>) => 
   dispatch({ type: ActionTypes.REMOVE_ASYNC_TASK });
 };
 
-/**
- * @summary Set repositories in redux store
- */
 export const setRepositories = () => async (dispatch: Dispatch<FetchAction>) => {
   dispatch({ type: ActionTypes.ADD_ASYNC_TASK });
   const repositories = await getRepositories();
@@ -129,71 +145,48 @@ export const setRepositories = () => async (dispatch: Dispatch<FetchAction>) => 
   dispatch({ type: ActionTypes.REMOVE_ASYNC_TASK });
 };
 
-/**
- * @summary Toggle full screen mode for extension
- */
-export const toggleFullScreenMode = () => async (dispatch: Dispatch<FetchAction>) => {
-  const newFullScreenModeState = await setFullScreenMode();
+export const setFullScreenMode = (isFullScreenMode: boolean) => async (dispatch: Dispatch<FetchAction>) => {
+  const newFullScreenModeState = await setFullScreenModeState(isFullScreenMode);
   dispatch({ type: ActionTypes.SET_FULL_SCREEN_MODE, payload: newFullScreenModeState });
 };
 
-// export const restoreSettings = () => async (dispatch: Dispatch<FetchAction>) => {
-//   const settings = await getSettings();
-//   dispatch({ type: ActionTypes.RESTORE_SETTINGS, payload: settings });
-//   dispatch({ type: ActionTypes.TRIGGER_SORT_DIRECTION });
-// };
+export const restoreSettings = () => async (dispatch: Dispatch<FetchAction>) => {
+  const settings = await getSettings();
+  await setFullScreenModeState(settings.isFullScreenMode);
+  dispatch({ type: ActionTypes.RESTORE_SETTINGS, payload: settings });
+  dispatch({ type: ActionTypes.TRIGGER_SORT_DIRECTION });
+};
 
-// export const clearSettings = () => async (dispatch: Dispatch<FetchAction>) => {
-//   const pristineState: SavedPrHubState = {
-//     settings: initialState.settings,
-//     ui: initialState.ui,
-//   };
-//   await setSettings(pristineState);
-//   dispatch({ type: ActionTypes.RESTORE_SETTINGS, payload: pristineState });
-//   dispatch({ type: ActionTypes.TRIGGER_SORT_DIRECTION });
-// };
+export const saveSettings = (defaultSettings: DefaultSettings) => async (dispatch: Dispatch<FetchAction>) => {
+  const settings = await setSettings(defaultSettings);
+  dispatch({ type: ActionTypes.RESTORE_SETTINGS, payload: settings });
+  dispatch({ type: ActionTypes.TRIGGER_SORT_DIRECTION });
+};
 
 export const onInitialLoad = () => {
   return (dispatch: Dispatch<any>) => {
-    // dispatch(restoreSettings());
     dispatch(setCurrentUser());
     dispatch(setRepositories());
     dispatch(setPullRequests());
+    dispatch(restoreSettings());
   };
 };
 
-// This all needs to be in context of setting a "default view"
-
-/**
- * @summary Get Azure Extension Storage Context
- */
 const getDataManagementContext = async (): Promise<IExtensionDataManager> => {
-  const extensionId = DevOps.getExtensionContext().id;
-  const accessToken = await DevOps.getAccessToken();
-  const extensionDataService = await DevOps.getService<IExtensionDataService>('ms.vss-features.extension-data-service');
+  const extensionId = getExtensionContext().id;
+  const accessToken = await getAccessToken();
+  const extensionDataService = await getService<IExtensionDataService>('ms.vss-features.extension-data-service');
   return extensionDataService.getExtensionDataManager(extensionId, accessToken);
 };
 
-export const getSettings = async (): Promise<SavedPrHubState> => {
+export const getSettings = async (): Promise<DefaultSettings> => {
   const dbKey = `prc-ext-data-default`;
   const context = await getDataManagementContext();
   return context.getValue(dbKey, { defaultValue: null });
 };
 
-export const setSettings = async (data: SavedPrHubState): Promise<SavedPrHubState> => {
+export const setSettings = async (data: DefaultSettings): Promise<DefaultSettings> => {
   const dbKey = `prc-ext-data-default`;
   const context = await getDataManagementContext();
-  const newSavedState: SavedPrHubState = {
-    settings: {
-      settingsLastSaved: new Date().toISOString(),
-      settingsPanelOpen: false,
-    },
-    ui: {
-      isFilterVisible: data.ui.isFilterVisible,
-      isFullScreenMode: data.ui.isFullScreenMode,
-      selectedTab: data.ui.selectedTab,
-      sortDirection: data.ui.sortDirection,
-    },
-  };
-  return context.setValue(dbKey, newSavedState);
+  return context.setValue(dbKey, data);
 };