+++ src/state/types.ts
@@ -4,12 +4,12 @@ import { IdentityRef } from 'azure-devops-extension-api/WebApi/WebApi';
 import { WorkItem } from 'azure-devops-extension-api/WorkItemTracking/WorkItemTracking';
 import { IdentityRefWithVote, PullRequestStatus, GitRepository } from 'azure-devops-extension-api/Git/Git';
 
-import { TabOptions } from '../tabs/TabTypes';
 import { ReviewerVoteNumber } from '../lib/enums';
 
 const GET_SETTINGS = 'getSettings';
 const SET_SETTINGS = 'setSettings';
 const ADD_ASYNC_TASK = 'addAsyncTask';
+const SET_FILTER_BAR = 'setFilterBar';
 const SET_CURRENT_USER = 'setCurrentUser';
 const SET_SELECTED_TAB = 'setSelectedTab';
 const RESTORE_SETTINGS = 'restoreSettings';
@@ -19,6 +19,7 @@ const SET_PULL_REQUESTS = 'setPullRequests';
 const REMOVE_ASYNC_TASK = 'removeAsyncTask';
 const TOGGLE_FILTER_BAR = 'toggleFilterBar';
 const DISPLAY_WORK_ITEMS = 'displayWorkItems';
+const SET_SORT_DIRECTION = 'setSortDirection';
 const SET_FULL_SCREEN_MODE = 'setFullScreenMode';
 const TOGGLE_SETTINGS_PANEL = 'toggleSettingsPanel';
 const REFRESH_PULL_REQUESTS = 'refreshPullRequests';
@@ -29,6 +30,7 @@ export const ActionTypes = {
   GET_SETTINGS,
   SET_SETTINGS,
   ADD_ASYNC_TASK,
+  SET_FILTER_BAR,
   SET_CURRENT_USER,
   SET_SELECTED_TAB,
   RESTORE_SETTINGS,
@@ -38,6 +40,7 @@ export const ActionTypes = {
   REMOVE_ASYNC_TASK,
   TOGGLE_FILTER_BAR,
   DISPLAY_WORK_ITEMS,
+  SET_SORT_DIRECTION,
   SET_FULL_SCREEN_MODE,
   TOGGLE_SETTINGS_PANEL,
   REFRESH_PULL_REQUESTS,
@@ -45,15 +48,25 @@ export const ActionTypes = {
   TRIGGER_SORT_DIRECTION,
 } as const;
 
+export type DefaultSettings = {
+  isFilterVisible: boolean;
+  isFullScreenMode: boolean;
+  selectedTab: TabOptions;
+  sortDirection: SortDirection;
+};
+
 export type Settings = {
   settingsLastSaved: string;
   settingsPanelOpen: boolean;
+  defaults: DefaultSettings;
 };
 
 export type PR = {
   pullRequestId: number;
   repositoryId: string;
   isDraft: boolean;
+  isActive: boolean;
+  isCompleted: boolean;
   isAutoComplete: boolean;
   hasMergeConflicts: boolean;
   status: PullRequestStatus;
@@ -82,11 +95,13 @@ export type Data = {
   asyncTaskCount: number;
 };
 
+export type SortDirection = 'desc' | 'asc';
+export type TabOptions = 'active' | 'draft' | 'recentlyCompleted';
 export type UI = {
   isFilterVisible: ObservableValue<boolean>;
   isFullScreenMode: boolean;
   selectedTab: TabOptions;
-  sortDirection: 'desc' | 'asc';
+  sortDirection: SortDirection;
 };
 
 export type PrHubState = {
@@ -94,5 +109,3 @@ export type PrHubState = {
   data: Data;
   ui: UI;
 };
-
-export type SavedPrHubState = Omit<PrHubState, 'data'>;