+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # [Unreleased]
 
+# 1.1.0 (December 31, 2019)
+
+✔ Set extension defaults with new settings experience.
+
+✔ Show 10 most recently completed PRs.
+
+- Filter by pull request author added to string filter.
+
+🐛 Reloading Pull Requests now preserving previously filtered values being applied.
+
 # 1.0.{2,3,4,5} (December 17, 2019)
 
 ✔ Fields including pull request id and target branch name are now searchable.